refactor: reservation 도메인 분리 & RoomEscape 명칭을 쓰고 있던 클래스 Reservation으로 변경
