+++ src/main/java/web/reservation/controller/ReservationController.java
@@ -1,37 +1,37 @@
-package web.controller;
+package web.reservation.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import web.dto.ReservationRequestDto;
-import web.dto.ReservationResponseDto;
-import web.exception.ReservationException;
-import web.service.RoomEscapeService;
+import web.reservation.dto.ReservationRequestDto;
+import web.reservation.dto.ReservationResponseDto;
+import web.reservation.exception.ReservationException;
+import web.reservation.service.ReservationService;
 
 import javax.validation.Valid;
 import java.net.URI;
 import java.time.LocalTime;
 
-import static web.exception.ErrorCode.NOT_UNIT_OF_30_MINUTES;
-import static web.exception.ErrorCode.OUT_OF_BUSINESS_HOURS;
+import static web.reservation.exception.ErrorCode.NOT_UNIT_OF_30_MINUTES;
+import static web.reservation.exception.ErrorCode.OUT_OF_BUSINESS_HOURS;
 
 @RequestMapping("/reservations")
 @RequiredArgsConstructor
 @RestController
-public class RoomEscapeController {
+public class ReservationController {
     public static final LocalTime BEGIN_TIME = LocalTime.of(11, 0, 0);
     public static final LocalTime LAST_TIME = LocalTime.of(20, 30, 0);
 
-    private final RoomEscapeService roomEscapeService;
+    private final ReservationService reservationService;
 
     @PostMapping
     public ResponseEntity<Void> reservation(@RequestBody @Valid ReservationRequestDto requestDto) {
         checkOutOfBusinessHours(requestDto.getTime());
         checkUnitOf30Minutes(requestDto.getTime());
 
-        long createdId = roomEscapeService.reservation(requestDto);
+        long createdId = reservationService.reservation(requestDto);
 
         return ResponseEntity.status(HttpStatus.CREATED)
                 .location(URI.create("/reservations/" + createdId))
@@ -52,7 +52,7 @@ public class RoomEscapeController {
 
     @GetMapping("/{reservationId}")
     public ResponseEntity<ReservationResponseDto> findReservationById(@PathVariable long reservationId) {
-        ReservationResponseDto responseDto = roomEscapeService.findReservationById(reservationId);
+        ReservationResponseDto responseDto = reservationService.findReservationById(reservationId);
         return ResponseEntity.ok()
                 .contentType(MediaType.APPLICATION_JSON)
                 .body(responseDto);
@@ -60,7 +60,7 @@ public class RoomEscapeController {
 
     @DeleteMapping("/{reservationId}")
     public ResponseEntity<Void> cancelReservation(@PathVariable long reservationId) {
-        roomEscapeService.cancelReservation(reservationId);
+        reservationService.cancelReservation(reservationId);
         return ResponseEntity.noContent()
                 .build();
     }
