+++ src/main/java/web/reservation/dto/ReservationRequestDto.java
@@ -1,4 +1,4 @@
-package web.dto;
+package web.reservation.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
